fix a link and update a link in the FLT blog post (#101)
The blog post was linking to a blueprint graph which no longer exists;
it was weird having the entire project represented by one very
disconnected and confusing and very-slowly-changing graph, so I switched
to "one graph per chapter" and this broke the blog post link.

I also updated the link to the Vantage seminar now that the video is
online.